Refactor hero section
