+++ src/styles/components/index.js
@@ -1,193 +1,6 @@
 import styled from "styled-components"
 import * as styleVars from "./variable"
 
-export const Hero = styled.section`
-  height: 100%;
-  position: relative;
-  padding: 0 2rem;
-  margin-bottom: 400px;
-
-  @media only screen and (min-width: ${styleVars.mdUp}) {
-    padding: 0 10rem;
-  }
-
-  @media only screen and (max-width: ${styleVars.smUp}) {
-    margin-bottom: 80px;
-  }
-
-  .hero__copy {
-    padding-top: 14rem;
-  }
-
-  .hero__heading {
-    text-align: left;
-    margin: auto;
-    font-size: 3.875em;
-    font-weight: normal;
-    color: ${styleVars.colorHackRed};
-
-    @media only screen and (max-width: ${styleVars.smUp}) {
-      text-align: center;
-      font-size: 4.8rem;
-    }
-  }
-
-  .hero__topsubheading {
-    font-size: 1.5em;
-    font-weight: 400;
-    text-align: left;
-    color: ${styleVars.colorHackRed};
-    margin-top: 0;
-    margin-bottom: 0;
-
-    @media only screen and (max-width: ${styleVars.smUp}) {
-      text-align: center;
-      padding-bottom: 0;
-    }
-  }
-
-  .hero__subheading {
-    font-size: 1.5em;
-    font-weight: 400;
-    text-align: left;
-    color: ${styleVars.hackBlack60};
-    margin-top: 0;
-    margin-bottom: 0;
-
-    @media only screen and (max-width: ${styleVars.smUp}) {
-      text-align: center;
-      padding-bottom: 0;
-    }
-  }
-
-  .hero__cta {
-    font-size: 1.4rem;
-    font-weight: 400;
-    text-align: center;
-
-    @media only screen and (min-width: ${styleVars.smUp}) {
-      font-size: 1.8rem;
-      text-align: left;
-    }
-  }
-
-  .hero__email {
-    font-weight: 400;
-    text-align: left;
-    margin-bottom: 1rem;
-    display: block;
-
-    @media only screen and (max-width: ${styleVars.smUp}) {
-      text-align: center;
-    }
-  }
-`
-
-export const HeroButton = styled.button`
-  //   font-family: Brown, sans-serif;
-  font-weight: 600;
-  font-size: 16px;
-  text-transform: none;
-  color: ${styleVars.colorHackRed};
-  border-radius: 0.5em;
-  border: 2px solid ${styleVars.colorHackRed};
-  transition: all 0.15s ease-in-out;
-  padding: 10px 32px;
-
-  &:hover {
-    background-color: ${styleVars.colorHackRed};
-    color: white;
-  }
-
-  @media only screen and (min-width: ${styleVars.smUp}) {
-    text-align: center;
-  }
-`
-
-export const ArrowButton = styled.button`
-  background-color: ${styleVars.colorHackRed};
-  color: #fff;
-  border-radius: 2em;
-  border: none;
-  transition: all 0.15s ease-in-out;
-  top: 1px;
-  right: 0;
-  height: 36px;
-  padding: 0 14px;
-  margin-left: -8px;
-
-  &:hover {
-    background-color: $color-hack-red-med;
-  }
-`
-
-export const HeroBackground = styled.img`
-  position: absolute;
-  height: auto;
-  object-fit: contain;
-  top: 100px;
-  z-index: -6;
-  width: 100%;
-
-  @media only screen and (max-width: ${styleVars.smUp}) {
-    position: relative;
-    max-width: 80%;
-    margin: 0;
-    padding: 0 2rem;
-  }
-`
-
-export const HeroForeground = styled.img`
-  position: absolute;
-  height: auto;
-  object-fit: contain;
-  z-index: -5;
-  top: 350px;
-  right: 0px;
-
-  @media only screen and (max-width: ${styleVars.smUp}) {
-    max-width: 90%;
-    margin: 0;
-    padding: 0 2rem;
-    padding-top: 7.3rem;
-    position: unset;
-    z-index: -1;
-  }
-`
-
-export const HeroBuildings = styled.img`
-  position: absolute;
-  height: auto;
-  object-fit: contain;
-  z-index: -4;
-  top: 290px;
-  right: 0px;
-
-  @media only screen and (max-width: ${styleVars.smUp}) {
-    max-width: 90%;
-    max-height: 80%;
-    margin-top: 120px;
-    padding: 0;
-    padding-top: 3rem;
-    z-index: 0;
-  }
-`
-
-export const HeroCord = styled.img`
-  position: absolute;
-  height: auto;
-  object-fit: contain;
-  z-index: -3;
-  top: 350px;
-  right: 0px;
-
-  @media only screen and (max-width: ${styleVars.smUp}) {
-    max-width: 100%;
-    margin: 0;
-    padding: 0;
-  }
-`
-
 export const SponsContent = styled.div`
   display: flex;
   padding: 2rem 1rem;
